libdecaf: Cleaned up to pass /W4 without warnings.

+++ src/libdecaf/src/cafe/libraries/gx2/gx2_state.cpp
@@ -187,7 +187,7 @@ GX2GetMiscParam(GX2MiscParam param)
       return sStateData->hangResetSwapsOutstanding;
       break;
    default:
-      return -1;
+      return static_cast<uint32_t>(-1);
    }
 }
